separated file AudioPlayer into a component and embedded it in Sources and Library
